services/invoice-parser-service/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
tabula-py==2.8.2

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# ML/Matching
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
sentence-transformers==2.2.2
numpy==1.26.2

# LLM Integration (Optional)
openai==1.3.7
langchain==0.0.340

# Database
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Events
aiokafka==0.9.0
redis==5.0.1

# Storage
minio==7.2.0
boto3==1.29.7

# Utils
python-multipart==0.0.6
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Monitoring
prometheus-client==0.19.0
